# openapi/v0.yaml
openapi: 3.1.0
info:
  title: Fluorite Registry API
  version: "0.1.0"
  description: >
    Registry for publishing and fetching compiled TurboWarp extensions built
    with the Fluorite Compiler. All endpoints are versioned under /v0.
servers:
  - url: https://fluorite.sdisk.us/v0
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Accepts two kinds of bearer token: session tokens (from /auth/login,
        full access, 7-day TTL) and automation tokens (from /auth/tokens,
        long-lived, restricted to scopes granted at creation — currently only
        "publish"). Scope enforcement happens server-side per-endpoint.
  parameters:
...
    cursorParam:
      name: cursor
      in: query
      schema:
        type: string
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
  responses:
    Error:
      description: Standard error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authenticated, but not permitted to perform this action.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: VERSION_PENDING_REVIEW
            message:
              type: string
              example: >-
                You have another version pending review. It needs to be
                approved or rejected before you can publish again.
            field:
              type: [string, "null"]
    User:
      type: object
      required: [namespace, displayName, type, trusted]
      properties:
        namespace:
          type: string
        displayName:
          type: string
        type:
          type: string
          enum: [admin, normal]
        trusted:
          type: boolean
          description: >-
            Whether this user's publishes are auto-approved (true after their
            first-ever publish has been reviewed and approved).
...
    AutomationToken:
      type: object
      required: [id, name, scopes, createdAt, lastUsedAt]
      properties:
        id:
          type: string
        name:
          type: string
          description: 'Caller-supplied label, e.g. "CI: fluorite-publish-action".'
        scopes:
          type: array
          items:
            type: string
            enum: [publish]
        createdAt:
          type: string
          format: date-time
        lastUsedAt:
          type: [string, "null"]
          format: date-time
...
paths:
  /auth/signup:
    post:
      tags: [Open]
      operationId: signup
      summary: Create an account. The first successful signup on a fresh registry is auto-promoted to admin.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [namespace, password]
              properties:
                namespace:
                  type: string
                password:
                  type: string
                  minLength: 8
                displayName:
                  type: string
      responses:
        "201":
          description: Account created.
          headers:
            X-Unread-Notifications:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
                  token:
                    type: string
        "409":
          $ref: "#/components/responses/Error"
  /auth/login:
    post:
      tags: [Open]
      operationId: login
      summary: Start a new session.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [namespace, password]
              properties:
                namespace:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Authenticated.
          headers:
            X-Unread-Notifications:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
                  token:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/Error"
...
  /auth/sessions/{id}:
    delete:
      tags: [Userspace]
      operationId: revokeSession
      summary: Revoke one specific session.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Session revoked.
        "404":
          $ref: "#/components/responses/NotFound"
  /auth/tokens:
    get:
      tags: [Userspace]
      operationId: listAutomationTokens
      summary: List the caller's automation tokens. Secrets are never returned after creation.
      responses:
        "200":
          description: Automation tokens.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AutomationToken"
    post:
      tags: [Userspace]
      operationId: createAutomationToken
      summary: >-
        Create a scoped, long-lived token for CI/automation (e.g.
        fluorite-publish-action). Unlike sessions, these don't expire on a
        timer — only on explicit revocation or a password change.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, scopes]
              properties:
                name:
                  type: string
                scopes:
                  type: array
                  items:
                    type: string
                    enum: [publish]
                  minItems: 1
      responses:
        "201":
          description: Token created. The plaintext token is returned once and cannot be retrieved again.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/AutomationToken"
                  - type: object
                    required: [token]
                    properties:
                      token:
                        type: string
  /auth/tokens/{id}:
    delete:
      tags: [Userspace]
      operationId: revokeAutomationToken
      summary: Revoke one automation token.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Revoked.
        "404":
          $ref: "#/components/responses/NotFound"
  /users:
    get:
      tags: [Open]
      operationId: listUsers
      summary: List users.
      security: []
      parameters:
        - $ref: "#/components/parameters/cursorParam"
        - $ref: "#/components/parameters/limitParam"
      responses:
        "200":
          description: A page of users.
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
                  nextCursor:
                    type: [string, "null"]
    post:
      tags: [Adminspace]
      operationId: createUser
      summary: Create an account directly, without self-signup.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [namespace, password]
              properties:
                namespace:
                  type: string
                password:
                  type: string
                displayName:
                  type: string
      responses:
        "201":
          description: Account created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Error"
...